Moved controller to jujuclient and controllerserver to controller.
