+++ worker/provisioner/provisioner_task.go
@@ -17,12 +17,12 @@ import (
 	"github.com/juju/juju/apiserver/params"
 	"github.com/juju/juju/cloudconfig/instancecfg"
 	"github.com/juju/juju/constraints"
+	"github.com/juju/juju/controller/authentication"
 	"github.com/juju/juju/environs"
 	"github.com/juju/juju/environs/config"
 	"github.com/juju/juju/environs/imagemetadata"
 	"github.com/juju/juju/environs/simplestreams"
 	"github.com/juju/juju/instance"
-	"github.com/juju/juju/jujuclient/authentication"
 	"github.com/juju/juju/network"
 	"github.com/juju/juju/storage"
 	coretools "github.com/juju/juju/tools"
